+++ service/service.go
@@ -2,7 +2,7 @@
  * @Author: LinkLeong link@icewhale.com
  * @Date: 2022-07-12 09:48:56
  * @LastEditors: LinkLeong
- * @LastEditTime: 2022-07-15 10:58:54
+ * @LastEditTime: 2022-07-27 10:28:48
  * @FilePath: /CasaOS/service/service.go
  * @Description:
  * @Website: https://www.casaos.io
@@ -33,31 +33,44 @@ type Repository interface {
 	Notify() NotifyServer
 	Rely() RelyService
 	System() SystemService
+	Shares() SharesService
+	Connections() ConnectionsService
 }
 
 func NewService(db *gorm.DB) Repository {
 	return &store{
-		app:    NewAppService(db),
-		user:   NewUserService(db),
-		docker: NewDockerService(),
-		casa:   NewCasaService(),
-		disk:   NewDiskService(db),
-		notify: NewNotifyService(db),
-		rely:   NewRelyService(db),
-		system: NewSystemService(),
+		app:         NewAppService(db),
+		user:        NewUserService(db),
+		docker:      NewDockerService(),
+		casa:        NewCasaService(),
+		disk:        NewDiskService(db),
+		notify:      NewNotifyService(db),
+		rely:        NewRelyService(db),
+		system:      NewSystemService(),
+		shares:      NewSharesService(db),
+		connections: NewConnectionsService(db),
 	}
 }
 
 type store struct {
-	db     *gorm.DB
-	app    AppService
-	user   UserService
-	docker DockerService
-	casa   CasaService
-	disk   DiskService
-	notify NotifyServer
-	rely   RelyService
-	system SystemService
+	db          *gorm.DB
+	app         AppService
+	user        UserService
+	docker      DockerService
+	casa        CasaService
+	disk        DiskService
+	notify      NotifyServer
+	rely        RelyService
+	system      SystemService
+	shares      SharesService
+	connections ConnectionsService
+}
+
+func (s *store) Connections() ConnectionsService {
+	return s.connections
+}
+func (s *store) Shares() SharesService {
+	return s.shares
 }
 
 func (c *store) Rely() RelyService {